Revert "[netstack3] Cleanup TCP established_send test"
Revert submission 663688

Reason for revert: accidentally submitted w/o going through CQ
Reverted Changes:
I48e37b26b:[netstack3] Don't ACK a pure ACK
Idafb09169:[netstack3] Cleanup TCP established_send test
I7148283f4:[netstack3] Prefer destructure over field access

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663932
Commit-Queue: Zeling Feng <zeling@google.com>

+++ src/connectivity/network/netstack3/core/src/transport/tcp/segment.rs
@@ -416,7 +416,7 @@ mod test {
         control: Option<Control>,
         data_len: u32,
         rcv_nxt: u32,
-        rcv_wnd: usize,
+        rcv_wnd: u32,
     }
     #[test_case(OverlapTestArgs{
         seg_seq: 1,
